move bid shares
