+++ as/cNFT/assembly/models/market.ts
@@ -9,11 +9,4 @@ export class Bid {
     currency: string
 }
 
-@nearBindgen
-export class BidShares {
-    public prev_owner: u16
-    public creator: u16
-    public owner: u16
-}
-
 export type BidsByBidder = Map<string, Bid>
