+++ as/cNFT/assembly/models/persistent_market.ts
@@ -1,13 +1,11 @@
 import { PersistentSet, PersistentUnorderedMap } from 'near-sdk-as'
 import { AccountId, TokenId } from '../types'
-import { Bid, Ask, BidShares, BidsByBidder } from './market'
+import { Bid, Ask, BidsByBidder } from './market'
 
 @nearBindgen
 export class PersistentMarket {
     private _tmap: PersistentUnorderedMap<TokenId, BidsByBidder>
     private _amap: PersistentUnorderedMap<AccountId, PersistentSet<TokenId>>
-    private _askmap: PersistentUnorderedMap<TokenId, Ask>
-    private _shmap: PersistentUnorderedMap<TokenId, BidShares>
 
     /**
      * @param prefix A prefix to use for every key of this map
@@ -20,14 +18,6 @@ export class PersistentMarket {
             AccountId,
             PersistentSet<TokenId>
         >('_amap' + prefix)
-
-        this._askmap = new PersistentUnorderedMap<TokenId, Ask>(
-            '_askmap' + prefix
-        )
-
-        this._shmap = new PersistentUnorderedMap<TokenId, BidShares>(
-            '_shmap' + prefix
-        )
     }
 
     /**
@@ -71,28 +61,7 @@ export class PersistentMarket {
         )
     }
 
-    /**
-     * Ask
-     */
-    set_ask(tokenId: TokenId, ask: Ask): void {
-        this._askmap.set(tokenId, ask)
-    }
-    get_ask(tokenId: TokenId): Ask {
-        return this._askmap.getSome(tokenId)
-    }
-    remove_ask(tokenId: TokenId): void {
-        this._askmap.delete(tokenId)
-    }
 
-    /**
-     * Bid shares
-     */
-    set_bid_shares(tokenId: TokenId, bidShares: BidShares): void {
-        this._shmap.set(tokenId, bidShares)
-    }
-    get_bid_shares(tokenId: TokenId): BidShares {
-        return this._shmap.getSome(tokenId)
-    }
 
     private _addToTokenBidMap(
         tokenId: TokenId,
