+++ as/cNFT/assembly/models/royalties.ts
@@ -1,0 +1,7 @@
+
+@nearBindgen
+export class BidShares {
+    public prev_owner: u16
+    public creator: u16
+    public owner: u16
+}
