+++ as/cNFT/assembly/utils/royalties.ts
@@ -1,5 +1,6 @@
 import { u128 } from 'near-sdk-as'
 import { Balance } from '../types'
+import { BidShares } from '../models/royalties'
 
 const ONE_HUNDRED_PERCENT: u16 = 10000
 
@@ -16,4 +17,15 @@ export function royalty_to_payout(
         u128.mul(amount_to_pay, u128.from(royalty_percentage)),
         u128.from(ONE_HUNDRED_PERCENT)
     )
+}
+
+export function split_share(percentage: u16, amount: u128): u128 {
+    return u128.div(
+        u128.mul(amount, u128.from(percentage)),
+        u128.from(ONE_HUNDRED_PERCENT)
+    )
+}
+
+export function calculate_owner_share(bidShares: BidShares): u16 {
+    return ONE_HUNDRED_PERCENT - bidShares.prev_owner - bidShares.creator
 }
